# openapi.json
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
  /agro-product/:
    get:
      operationId: agro_product_retrieve
      description: |-
        🛍️ Barcha turdagi dorilar uchun API
        ---
        - **GET** `/agro-product/` - Barcha mahsulotlarni olish
      tags:
      - agro-product
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /agro-product/{id}/:
    get:
      operationId: agro_product_retrieve_2
      description: |-
        🛍️ Bitta dori uchun API
        ---
        - **GET** `/agro-product/1/` - Bitta mahsulotlarni olish
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - agro-product
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /agro-product/categories/:
    get:
      operationId: agro_product_categories_retrieve
      description: |-
        🛍️ Category API
        ---
        - **GET** `/agro-product/categories/` - Barcha categoriyalarni olish
      tags:
      - agro-product
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /agro-product/drugs/:
    get:
      operationId: agro_product_drugs_retrieve
      description: |-
        🛍️ Statusi Dori bo'lgan mavsulot uchun API
        ---
        - **GET** `/agro-product/drugs/` - Barcha mahsulotlarni olish
      tags:
      - agro-product
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /agro-product/fertilizers/:
    get:
      operationId: agro_product_fertilizers_retrieve
      description: |-
        🛍️ Statusi O'g'it bo'lgan mahsulotlar uchun API
        ---
        - **GET** `/agro-product/fertilizers/` - Barcha mahsulotlarni olish
      tags:
      - agro-product
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/schema/:
    get:
      operationId: api_schema_retrieve
      description: |-
        OpenApi3 schema for this API. Format can be selected via content negotiation.

        - YAML: application/vnd.oai.openapi
        - JSON: application/vnd.oai.openapi+json
      parameters:
      - in: query
        name: format
        schema:
          type: string
          enum:
          - json
          - yaml
      - in: query
        name: lang
        schema:
          type: string
          enum:
          - af
          - ar
          - ar-dz
          - ast
          - az
          - be
          - bg
          - bn
          - br
          - bs
          - ca
          - ckb
          - cs
          - cy
          - da
          - de
          - dsb
          - el
          - en
          - en-au
          - en-gb
          - eo
          - es
          - es-ar
          - es-co
          - es-mx
          - es-ni
          - es-ve
          - et
          - eu
          - fa
          - fi
          - fr
          - fy
          - ga
          - gd
          - gl
          - he
          - hi
          - hr
          - hsb
          - hu
          - hy
          - ia
          - id
          - ig
          - io
          - is
          - it
          - ja
          - ka
          - kab
          - kk
          - km
          - kn
          - ko
          - ky
          - lb
          - lt
          - lv
          - mk
          - ml
          - mn
          - mr
          - ms
          - my
          - nb
          - ne
          - nl
          - nn
          - os
          - pa
          - pl
          - pt
          - pt-br
          - ro
          - ru
          - sk
          - sl
          - sq
          - sr
          - sr-latn
          - sv
          - sw
          - ta
          - te
          - tg
          - th
          - tk
          - tr
          - tt
          - udm
          - ug
          - uk
          - ur
          - uz
          - vi
          - zh-hans
          - zh-hant
      tags:
      - api
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/vnd.oai.openapi:
              schema:
                type: object
                additionalProperties: {}
            application/yaml:
              schema:
                type: object
                additionalProperties: {}
            application/vnd.oai.openapi+json:
              schema:
                type: object
                additionalProperties: {}
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid